Added expense classes, including DefaultExpenses as an abstract class inherited by other classes such as InstallmentExpenses, Expenses (general expenses including variable ones), and FixedExpenses. Created Category class to reference categories for each expense.
Focused on InstallmentExpenses, creating repository, service, and controller classes, along with a DTO for handling HTTP data to create installment expenses. The createInstallmentExpense method in the service adds an entry to the installment_expenses table and generates entries for each installment in the expenses table, requiring the creation of an expenses repository.